Android.mk: Use SYSLINUX as INSTALLED_BOOTLOADER_MODULE
Since we will be using standard Android bootimages for the various
kernels, we need to create a partition which just contains the
SYSLINUX binaries and configuration.

This logic used to be handled by SYSLINUX hacks to Android build/,
but we can now leverage the normal mechanisms for creating the
bootloader image.

Change-Id: Ica78c666ab22940ba09cc923dedb2b063c1d96b9
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>