Explicitly define rule for bootloader
The rule for bootloader relied on a symbol that was set in
build/target/board/Android.mk, which set up a dependency between this
Android.mk on that one.  Since there is no guaranteed order, nor any way
to express a dependency between makefiles at this level, this dependency
must be removed.

There are only two possible values for INSTALLED_BOOTLOADER_MODULE, null
or $(PRODUCT_OUT)/bootloader and this is set depending on whether or not there
is a bootloader.  This particular project will only build provided there is a
bootloader, which implies that we can safely replace INSTALLED_BOOTLOADER_MODULE
with $(PRODUCT_OUT)/bootloader.  This is made more explicit by ensuring that
the Android.mk here checks TARGET_NO_BOOTLOADER as well as TARGET_USE_SYSLINUX.

Change-Id: Ie82d4c9a38d8190c4aa902374c4dcec61122396c