Increase the syslinux image size by a custom value
Added support for custom platforms to define a custom variable in
their BoardConfig.mk, called TARGET_SYSLINUX_IMAGE_EXTRA_SPACE, which
will cause increasing the FAT syslinux image size by exactly the value
of this variable. The android-image.sh also got a new optional
parameter, defining this extra size value.

The optional parameter is on the same line as SYSLINUX_MK_IMG, to
avoid expanding an ugly empty line when
TARGET_SYSLINUX_IMAGE_EXTRA_SPACE is not defined.

Will be needed by a subsequent change in GreenRidge configuration, for
the syslinux image to be able to accomodate FW binaries for BIOS and
EC.

Change-Id: I874036c732a0eabb1c5b084e3237dafdd13e06dc
Signed-off-by: Stefan A. Popa <stefan.a.popa@intel.com>
Signed-off-by: Alin Dobre <alin.m.dobre@intel.com>